apps/ingest/requirements.txt
requests==2.32.3
feedparser==6.0.11
python-dateutil==2.9.0.post0
redis==5.0.6
kafka-python-ng[crc32c,lz4,snappy,zstd]==2.2.3
pydantic==2.7.4
certifi==2024.8.30
backoff==2.2.1
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
trafilatura==2.0.0